requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
attrs==25.1.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
distlib==0.3.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
execnet==2.1.1
fastapi==0.115.8
filelock==3.10.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
platformdirs==3.1.1
pluggy==1.5.0
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.23.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.21.0
watchfiles==1.0.4
websockets==15.0
